migrate relevant tests
