+++ marconi-cardano-core/test/Spec.hs
@@ -2,6 +2,8 @@
 
 module Main (main) where
 
+import Spec.Marconi.Cardano.Core.Logger qualified as Logger
+import Spec.Marconi.Cardano.Core.Orphans qualified as Orphans
 import Test.Tasty (TestTree, defaultMain, testGroup)
 
 main :: IO ()
@@ -12,4 +14,4 @@ tests :: TestTree
 tests =
   testGroup
     "marconi-cardano-core"
-    []
+    [Logger.tests, Orphans.tests]
