+++ marconi-cardano-core/test/Spec/Marconi/Cardano/Core/Logger/Golden/start-from-later-point-logging.txt
@@ -1,0 +1,7 @@
+Starting from ChainPoint(Slot 50000, BlockHash 6161616161616161616161616161616161616161616161616161616161616161). 
+Synchronising (80.00%). Current synced point is ChainPoint(Slot 80000, BlockHash 6161616161616161616161616161616161616161616161616161616161616161) and current node tip is ChainTip(Slot 100000, BlockHash 6161616161616161616161616161616161616161616161616161616161616161, BlockNo 100000). Processed 30 blocks and 0 rollbacks in the last 10s (3 blocks/s).
+Synchronising (90.00%). Current synced point is ChainPoint(Slot 90000, BlockHash 6161616161616161616161616161616161616161616161616161616161616161) and current node tip is ChainTip(Slot 100000, BlockHash 6161616161616161616161616161616161616161616161616161616161616161, BlockNo 100000). Processed 10 blocks and 0 rollbacks in the last 10s (1 blocks/s).
+Synchronising (95.00%). Current synced point is ChainPoint(Slot 95000, BlockHash 6161616161616161616161616161616161616161616161616161616161616161) and current node tip is ChainTip(Slot 100000, BlockHash 6161616161616161616161616161616161616161616161616161616161616161, BlockNo 100000). Processed 5 blocks and 2 rollbacks in the last 10s (0 blocks/s).
+Synchronising (almost synced). Current synced point is ChainPoint(Slot 99995, BlockHash 6161616161616161616161616161616161616161616161616161616161616161) and current node tip is ChainTip(Slot 100000, BlockHash 6161616161616161616161616161616161616161616161616161616161616161, BlockNo 100000). Processed 5 blocks and 1 rollbacks in the last 10s (0 blocks/s).
+Synchronising (almost synced). Current synced point is ChainPoint(Slot 99999, BlockHash 6161616161616161616161616161616161616161616161616161616161616161) and current node tip is ChainTip(Slot 100000, BlockHash 6161616161616161616161616161616161616161616161616161616161616161, BlockNo 100000). Processed 5 blocks and 1 rollbacks in the last 10s (0 blocks/s).
+Fully synchronised. Current synced point is ChainPoint(Slot 100000, BlockHash 6161616161616161616161616161616161616161616161616161616161616161) and current node tip is ChainTip(Slot 100000, BlockHash 6161616161616161616161616161616161616161616161616161616161616161, BlockNo 100000). Processed 5 blocks and 3 rollbacks in the last 10s.
