+++ marconi-cardano-core/test/Spec/Marconi/ChainIndex/Logger/Golden/start-from-genesis-logging.txt
