+++ marconi-cardano-core/test/Spec/Marconi/ChainIndex/Logger/Golden/start-from-later-point-logging.txt
