+++ marconi-chain-index/test/Spec/Marconi/ChainIndex/Indexers/UtxoQuery.hs
@@ -37,8 +37,8 @@ import Spec.Marconi.ChainIndex.Indexers.Datum qualified as Test.Datum
 import Spec.Marconi.ChainIndex.Indexers.Spent qualified as Test.Spent
 import Spec.Marconi.ChainIndex.Indexers.Utxo qualified as Test.Utxo
 import Test.Gen.Cardano.Api.Typed qualified as CGen
-import Test.Gen.Marconi.ChainIndex.Mockchain qualified as Gen
-import Test.Gen.Marconi.ChainIndex.Types qualified as CGen
+import Test.Gen.Marconi.Cardano.Core.Mockchain qualified as Gen
+import Test.Gen.Marconi.Cardano.Core.Types qualified as CGen
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.Hedgehog (testPropertyNamed)
